Fixing bugs in emap, adding multioutput network representation and writer for emap, adding an adder extraction algorithm, adding new standard cell view for multioutput gates, adding utils to decompose multioutput cells into single output for verification reasons, and adding test cases

+++ include/mockturtle/utils/include/supergate.hpp
@@ -77,7 +77,7 @@ struct supergate
   composed_gate<NInputs> const* root{};
 
   /* area */
-  float area{ 0.0 };
+  double area{ 0.0 };
 
   /* pin-to-pin delay */
   std::array<float, NInputs> tdelay{};
